Move TypeNameParser to a separate file.
