+++ mcs/class/System.Private.CoreLib/System.Private.CoreLib.csproj
@@ -126,6 +126,7 @@
       <Compile Include="System\Nullable.cs" />
       <Compile Include="System\String.cs" />
       <Compile Include="System\Type.cs" />
+      <Compile Include="System\TypeNameParser.cs" />
       <Compile Include="System\__ComObject.cs" />
       <Compile Include="System.Globalization\GlobalizationMode.cs" />
       <Compile Include="System.Globalization\CultureData.cs" />
